+++ Packages/CodeEditTextView/Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.3
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "CodeEditTextView",
-    platforms: [.macOS(.v13)],
+    platforms: [.macOS(.v26)],
     products: [
         // A Fast, Efficient text view for code.
         .library(