+++ Packages/WelcomeWindow/Package.swift
@@ -1,11 +1,11 @@
-// swift-tools-version: 6.0.0
+// swift-tools-version: 6.3
 
 import PackageDescription
 
 let package = Package(
     name: "WelcomeWindow",
     platforms: [
-        .macOS(.v13)
+        .macOS(.v26)
     ],
     products: [
         .library(